Балансировка, TODO в рекомендациях, врач теперь может указать желаемое время начала смен
(cherry picked from commit 53eb226118fc589bf82095ed95c15977b1d44f92)

+++ src/Entity/TempDoctorSchedule.php
@@ -33,8 +33,8 @@ class TempDoctorSchedule
     #[ORM\Column(type: 'integer', nullable: true)]
     private ?int $offMinutes = null;
 
-    #[ORM\Column(type: 'integer', nullable: true)]
-    private ?int $workHours = null;
+    #[ORM\Column(type: 'float', nullable: true)]
+    private ?float $workHours = null;
 
     #[ORM\Column(type: 'integer', nullable: true, options: ["comment" => "Планируемое количество исследований"])]
     private ?int $studyCount = null;
@@ -111,12 +111,12 @@ class TempDoctorSchedule
         return $this;
     }
 
-    public function getWorkHours(): ?int
+    public function getWorkHours(): ?float
     {
         return $this->workHours;
     }
 
-    public function setWorkHours(?int $workHours): TempDoctorSchedule
+    public function setWorkHours(?float $workHours): TempDoctorSchedule
     {
         $this->workHours = $workHours;
         return $this;
